Changed ints to long long, 5/5

+++ Contests/Euler+/002/sumEvenFib.cpp
@@ -24,9 +24,9 @@
 using namespace std;
 
 
-vector <int> fib(long n);
+vector <long> fib(long n);
 //void printVec(const vector<int>& v);
-int evenFibSum (const vector<int>& fib, int n);
+long long evenFibSum (const vector<long>& fib, long n);
 
 int main(){
     int t;
@@ -34,9 +34,9 @@ int main(){
     for(int a0 = 0; a0 < t; a0++){
         long n;
         cin >> n;
-        vector<int> fibonacciNums = fib(n);
+        vector<long> fibonacciNums = fib(n);
         //printVec(fibonacciNums);
-        int sum = evenFibSum(fibonacciNums, n);
+        long long sum = evenFibSum(fibonacciNums, n);
         cout << sum << endl;
     }
 
@@ -46,35 +46,35 @@ int main(){
 }
 
 
-vector <int> fib(long n){
-	vector <int> fibNums;
-	fibNums.push_back(1);
-	fibNums.push_back(2);
-	long sum;
-	for(int i = 2; fibNums[i-1] <= n; i++){
+vector <long> fib(long n){
+    vector <long> fibNums;
+    fibNums.push_back(1);
+    fibNums.push_back(2);
+    long sum;
+    for(int i = 2; fibNums[i-1] <= n; i++){
 
-		sum = fibNums[i-1] + fibNums[i-2];
-		fibNums.push_back(sum);
-		}
-	
-	return fibNums;
+        sum = fibNums[i-1] + fibNums[i-2];
+        fibNums.push_back(sum);
+        }
+    
+    return fibNums;
 
 }
 
 // void printVec(const vector<int>& v){
-// 	cout << "in print" << endl;
-// 	for(int i = 0; i < v.size(); i++){
-// 		cout << v[i] << "\t";
-// 	}
-// 	cout << endl;
+//     cout << "in print" << endl;
+//     for(int i = 0; i < v.size(); i++){
+//         cout << v[i] << "\t";
+//     }
+//     cout << endl;
 // }
 
-int evenFibSum (const vector<int>& fib, int n){
-	int evenSum = 0;
-	for(int i = 0; i < fib.size(); i++){
-		if(fib[i] <= n && fib[i]%2==0){
-			evenSum += fib[i];
-		}
-	}
-	return evenSum;
+long long evenFibSum (const vector<long>& fib, long n){
+    long long evenSum = 0;
+    for(int i = 0; i < fib.size(); i++){
+        if(fib[i] <= n && fib[i]%2==0){
+            evenSum += fib[i];
+        }
+    }
+    return evenSum;
 }
